pyproject.toml
[tool.poetry]
name = "pentulz-backend"
version = "0.1.0"
description = ""
authors = ["Pentulz Team"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.116.1"
uvicorn = {version="^0.35.0", extras=["standard"]}
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.43"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
alembic = "^1.16.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
anyio = "^4.10.0"
pylint = "^3.3.8"
black = "^25.1.0"
autoflake = "^2.3.1"
isort = "^6.0.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=app --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ["py310"]
